asdf

+++ P2/src/option1b.c
@@ -106,7 +106,7 @@ int main(int argc, char **argv) {
         }
     }
 
-    int *displs = malloc(sizeof(int) * size);
+    int *displs = malloc(sizeof(int) * (size_t)size);
 
     for (int i = 0; i < size; i++)
         displs[i] = p[i];
